Refactor validations under `validate` command
