+++ .travis.yml
@@ -43,17 +43,19 @@ jobs:
   include:
     - stage: test_changed
       script:
-        - ddev dep verify
-        - ddev manifest verify --include-extras
-        - ddev metadata verify
+        - ddev validate dep
+        - ddev validate manifest --include-extras
+        - ddev validate metadata
+        - ddev validate service-checks
         - ddev test --cov
         - ddev test --bench || true
     - stage: test
       env: CHECK=datadog_checks_base PYTHON3=true
       script:
-        - ddev dep verify
-        - ddev manifest verify --include-extras
-        - ddev metadata verify
+        - ddev validate dep
+        - ddev validate manifest --include-extras
+        - ddev validate metadata
+        - ddev validate service-checks
         - travis_retry ddev test --cov ${CHECK}
         - ddev test ${CHECK} --bench || true
     - stage: test
