+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/__init__.py
@@ -6,13 +6,11 @@ from .config import config
 from .create import create
 from .dep import dep
 from .env import env
-from .manifest import manifest
 from .meta import meta
-from .metadata import metadata
 from .release import release
 from .run import run
-from .service_checks import service_checks
 from .test import test
+from .validate import validate
 
 ALL_COMMANDS = (
     clean,
@@ -20,11 +18,9 @@ ALL_COMMANDS = (
     create,
     dep,
     env,
-    manifest,
     meta,
-    metadata,
     release,
     run,
-    service_checks,
     test,
+    validate,
 )
